requirements.txt
Flask==3.0.0
feedparser==6.0.11
requests==2.31.0
lxml==5.0.0
beautifulsoup4==4.12.3
sentry-sdk[flask]==1.43.0
opencv-python-headless==4.10.0.84
ollama==0.3.3
pillow==11.0.0